requirements.txt
anyio==3.6.2
attrs==22.2.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
fastapi==0.95.1
h11==0.14.0
idna==3.4
npm==0.1.1
optional-django==0.1.0
packaging==23.0
paho-mqtt==1.6.1
ply==3.11
pycparser==2.21
pydantic==1.10.7
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
